# fix pulping being too fast