#### Summary
None
#### Purpose of change
#30253 was a substantial change, but apparently not all stuff that needed update was updated. Apparently pulping speed was left unedited, which shifted then long action to a laughing few seconds
#### Describe the solution
Extend the duration of single corpse punch to 6 seconds instead of 1 second
#### Describe alternatives you've considered
Some another number longer than 1 second
#### Testing
no weapon, default survivor with str 8
| survival level | time before change | time after change |
| -------------- | ------------------ | ----------------- |
| 0               | 8 seconds                | 48 seconds       |
| 5               | 6 seconds                | 36 seconds       |
| 10             | 4 seconds                | 24 seconds       |

the timing is actually pretty inconsistent, there can be cases where you smash the corpse for a 30 seconds, and the second later you smash it for 50 seconds; no real corellation found, but on average it seems to correspond to this table