@GuardianDll 2025-01-09
Bash or cut damage of wielded weapon, strength, and corpse being damaged all contribute to pulping speed, yes - its a preexisting code that was somewhat obsoleted because of how fast pulping was
Any butchery action, including dismemberment, is not changed in this PR, tho i guess there is more reason to dismember monsters now, for some characters it may be faster than smashing