@GuardianDll 2025-01-09
there are plenty of factors, they are all right above my edit - skill, monster volume, your stamina etc